tools/serveurRDFGraphQLV2/graphql_sparql_api/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
graphene==3.3
SPARQLWrapper==2.0.0
authlib==1.2.1
pyjwt==2.8.0
python-dotenv==1.0.0
unidecode==1.3.6
requests==2.31.0
gunicorn==21.2.0

# Tests
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-flask==1.2.0

# Dev tools
black==23.9.1
flake8==6.1.0
